feat: implement channel cleanup, leave
