+++ backend/bin/client/src/lib.rs
@@ -22,7 +22,7 @@ use anyhow::{anyhow, Context};
 use futures::{future::poll_fn, ready};
 use headless_talk::{
     config::ClientEnv,
-    updater::{Credential, TalkInitializer},
+    initializer::{Credential, TalkInitializer},
     ClientStatus, HeadlessTalk,
 };
 use talk_loco_client::futures_loco_protocol::LocoClient;
